Move memory tracker to Execution Context
