+++ src/graph/context/QueryContext.h
@@ -22,9 +22,6 @@
 #include "parser/SequentialSentences.h"
 
 namespace nebula {
-
-class MemoryTracker;
-
 namespace graph {
 
 /***************************************************************************
@@ -52,7 +49,7 @@ class QueryContext {
                meta::MetaClient* metaClient,
                CharsetInfo* charsetInfo);
 
-  virtual ~QueryContext() = default;
+  virtual ~QueryContext();
 
   void setRCtx(RequestContextPtr rctx) { rctx_ = std::move(rctx); }
 
@@ -89,7 +86,6 @@ class QueryContext {
   int64_t genId() const { return idGen_->id(); }
 
   SymbolTable* symTable() const { return symTable_.get(); }
-  MemoryTracker* memTracker() const { return memTracker_.get(); }
 
   void setPartialSuccess() {
     DCHECK(rctx_ != nullptr);
@@ -118,7 +114,6 @@ class QueryContext {
   std::unique_ptr<ObjectPool> objPool_;
   std::unique_ptr<IdGenerator> idGen_;
   std::unique_ptr<SymbolTable> symTable_;
-  std::unique_ptr<MemoryTracker> memTracker_;
 
   std::atomic<bool> killed_{false};
 };
