+++ src/graph/executor/test/CMakeLists.txt
@@ -46,6 +46,7 @@ SET(EXEC_QUERY_TEST_OBJS
     $<TARGET_OBJECTS:util_obj>
     $<TARGET_OBJECTS:idgenerator_obj>
     $<TARGET_OBJECTS:graph_context_obj>
+    $<TARGET_OBJECTS:runtime_obj>
     $<TARGET_OBJECTS:graph_auth_obj>
     $<TARGET_OBJECTS:expr_visitor_obj>
     $<TARGET_OBJECTS:graph_obj>
