+++ src/graph/planner/test/CMakeLists.txt
@@ -49,6 +49,7 @@ nebula_add_test(
         $<TARGET_OBJECTS:idgenerator_obj>
         $<TARGET_OBJECTS:graph_context_obj>
         $<TARGET_OBJECTS:memory_obj>
+        $<TARGET_OBJECTS:runtime_obj>
     LIBRARIES
         gtest
         ${PROXYGEN_LIBRARIES}
