+++ src/graph/util/test/CMakeLists.txt
@@ -43,8 +43,8 @@ nebula_add_test(
         $<TARGET_OBJECTS:parser_obj>
         $<TARGET_OBJECTS:graph_context_obj>
         $<TARGET_OBJECTS:memory_obj>
+        $<TARGET_OBJECTS:runtime_obj>
         $<TARGET_OBJECTS:validator_obj>
-        $<TARGET_OBJECTS:version_obj>
     LIBRARIES
         gtest
         gtest_main
