+++ src/graph/validator/test/CMakeLists.txt
@@ -18,6 +18,7 @@ set(VALIDATOR_TEST_LIBS
     $<TARGET_OBJECTS:parser_obj>
     $<TARGET_OBJECTS:idgenerator_obj>
     $<TARGET_OBJECTS:graph_context_obj>
+    $<TARGET_OBJECTS:runtime_obj>
     $<TARGET_OBJECTS:graph_auth_obj>
     $<TARGET_OBJECTS:graph_session_obj>
     $<TARGET_OBJECTS:expression_obj>
