+++ src/graph/visitor/test/CMakeLists.txt
@@ -24,6 +24,7 @@ nebula_add_test(
         $<TARGET_OBJECTS:parser_obj>
         $<TARGET_OBJECTS:idgenerator_obj>
         $<TARGET_OBJECTS:graph_context_obj>
+        $<TARGET_OBJECTS:runtime_obj>
         $<TARGET_OBJECTS:graph_auth_obj>
         $<TARGET_OBJECTS:expression_obj>
         $<TARGET_OBJECTS:network_obj>
