+++ src/parser/test/CMakeLists.txt
@@ -38,6 +38,7 @@ set(PARSER_TEST_LIBS
     $<TARGET_OBJECTS:util_obj>
     $<TARGET_OBJECTS:expr_visitor_obj>
     $<TARGET_OBJECTS:graph_context_obj>
+    $<TARGET_OBJECTS:runtime_obj>
     $<TARGET_OBJECTS:planner_obj>
     $<TARGET_OBJECTS:validator_obj>
     $<TARGET_OBJECTS:idgenerator_obj>
